Refactor a function in main and a bit in settings
