+++ app/src/main/java/ch/epfl/sweng/studyup/settings/SettingsActivity.java
@@ -28,9 +28,7 @@ import ch.epfl.sweng.studyup.teacher.QuestsActivityTeacher;
 import ch.epfl.sweng.studyup.utils.Constants;
 import ch.epfl.sweng.studyup.utils.GlobalAccessVariables;
 import ch.epfl.sweng.studyup.utils.RefreshContext;
-import ch.epfl.sweng.studyup.utils.Utils;
 
-import static ch.epfl.sweng.studyup.player.HomeActivity.clearCacheToLogOut;
 import static ch.epfl.sweng.studyup.utils.Constants.LANG_SETTINGS_KEYWORD;
 import static ch.epfl.sweng.studyup.utils.Constants.PERSIST_LOGIN_FILENAME;
 import static ch.epfl.sweng.studyup.utils.Constants.USER_PREFS;
@@ -43,15 +41,10 @@ public class SettingsActivity extends RefreshContext {
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_settings);
-        setCheckBoxNPC();
-    }
-
-    @Override
-    protected void onResume() {
-        super.onResume();
         if(Player.get().isTeacher()) {
             ((TextView) findViewById(R.id.textCourseButton)).setText(R.string.course_choice_settings_button_teacher);
         }
+        setCheckBoxNPC();
     }
 
     private void setCheckBoxNPC() {
@@ -146,7 +139,16 @@ public class SettingsActivity extends RefreshContext {
         GlobalAccessVariables.NPCInteractionState = !GlobalAccessVariables.NPCInteractionState;
     }
 
-        public void onBackButton(View view) {
+    public static void clearCacheToLogOut(Context context) {
+        FileCacher<String[]> persistLogin = new FileCacher<>(context, PERSIST_LOGIN_FILENAME);
+        try {
+            persistLogin.clearCache();
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+    }
+
+    public void onBackButton(View view) {
         finish();
         overridePendingTransition(android.R.anim.fade_in, android.R.anim.fade_out);
     }
